[EI-136] [api] rename: type -> types
